refactor(meta): align table DDL event conventions
