+++ config/metasrv.example.toml
@@ -314,7 +314,9 @@ create_topic_timeout = "30s"
 ttl = "90d"
 ## Event types to record. Current available event types: `region_migration`,
 ## `create_database`, `alter_database`, `drop_database`, `create_flow`,
-## `drop_flow`, `create_view`, `drop_view`, `repartition`,
+## `drop_flow`, `create_table`, `create_logical_tables`, `alter_table`,
+## `alter_logical_tables`, `drop_table`, `undrop_table`, `purge_dropped_table`,
+## `truncate_table`, `create_view`, `drop_view`, `repartition`,
 ## `repartition_group`.
 ## When omitted, all current and future event types are recorded.
 ## Set to an empty array to disable event recording.
