+++ config/standalone.example.toml
@@ -905,6 +905,23 @@ default_ratio = 1.0
 ttl = "90d"
 ## Event types to record. Current available event types: `create_database`,
 ## `alter_database`, `drop_database`, `create_flow`, `drop_flow`,
+## `create_table`, `create_logical_tables`, `alter_table`, `alter_logical_tables`,
+## `drop_table`, `undrop_table`, `purge_dropped_table`, `truncate_table`,
+## `create_view`, `drop_view`.
+## When omitted, all current and future event types are recorded.
+## Set to an empty array to disable event recording.
+#+ event_types = ["create_database"]
+
+## The memory options.
+[memory]
+## Whether to enable heap profiling activation during startup.
+## When enabled, heap profiling will be activated if the `MALLOC_CONF` environment variable
+## is set to "prof:true,prof_active:false". The official image adds this env variable.
+## Default is true.
+enable_heap_profiling = true## Event types to record. Current available event types: `create_database`,
+## `alter_database`, `drop_database`, `create_flow`, `drop_flow`,
+## `create_table`, `create_logical_tables`, `alter_table`, `alter_logical_tables`,
+## `drop_table`, `undrop_table`, `purge_dropped_table`, `truncate_table`,
 ## `create_view`, `drop_view`.
 ## When omitted, all current and future event types are recorded.
 ## Set to an empty array to disable event recording.
