+++ src/common/meta/src/ddl/tests/alter_logical_tables.rs
@@ -13,21 +13,15 @@
 // limitations under the License.
 
 use std::assert_matches;
-use std::collections::HashSet;
 use std::sync::Arc;
 
 use api::region::RegionResponse;
 use api::v1::meta::Peer;
 use api::v1::region::sync_request::ManifestInfo;
 use api::v1::region::{MetricManifestInfo, RegionRequest, SyncRequest, region_request};
-use api::v1::value::ValueData;
-use api::v1::{ColumnDataType, SemanticType, Value};
+use api::v1::{ColumnDataType, SemanticType};
 use common_catalog::consts::{DEFAULT_CATALOG_NAME, DEFAULT_SCHEMA_NAME};
-use common_event_recorder::EventTypeFilter;
-use common_procedure::{
-    ChildSubmissionOutcome, EventContext, EventTrigger, Procedure, ProcedureId, ProcedureState,
-    RetryPhase, Status,
-};
+use common_procedure::{Procedure, ProcedureId, Status};
 use common_procedure_test::MockContextProvider;
 use store_api::metadata::ColumnMetadata;
 use store_api::metric_engine_consts::{ALTER_PHYSICAL_EXTENSION_KEY, MANIFEST_INFO_EXTENSION_KEY};
@@ -37,7 +31,6 @@ use store_api::storage::consts::ReservedColumnId;
 use tokio::sync::mpsc;
 
 use crate::ddl::alter_logical_tables::AlterLogicalTablesProcedure;
-use crate::ddl::event::table::TABLE_DDL_PAYLOAD_VERSION;
 use crate::ddl::test_util::alter_table::TestAlterTableExprBuilder;
 use crate::ddl::test_util::columns::TestColumnDefBuilder;
 use crate::ddl::test_util::datanode_handler::DatanodeWatcher;
@@ -54,7 +47,7 @@ use crate::rpc::ddl::AlterTableTask;
 use crate::rpc::router::{Region, RegionRoute};
 use crate::test_util::{MockDatanodeManager, new_ddl_context};
 
-fn make_alter_logical_table_add_column_task(
+pub(crate) fn make_alter_logical_table_add_column_task(
     schema: Option<&str>,
     table: &str,
     add_columns: Vec<String>,
@@ -148,124 +141,6 @@ fn assert_alters_request(
             *region_id,
             "actual region id: {}",
             RegionId::from_u64(req.requests[i].region_id)
-        );
-    }
-}
-
-#[test]
-fn test_submitted_event_has_one_row_per_logical_table_and_full_payload() {
-    let node_manager = Arc::new(MockDatanodeManager::new(()));
-    let ddl_context = new_ddl_context(node_manager);
-    let tasks = vec![
-        make_alter_logical_table_add_column_task(
-            Some(DEFAULT_SCHEMA_NAME),
-            "table1",
-            vec!["column1".to_string()],
-        ),
-        make_alter_logical_table_add_column_task(
-            Some(DEFAULT_SCHEMA_NAME),
-            "table2",
-            vec!["column2".to_string()],
-        ),
-    ];
-    let expected_tasks = serde_json::to_value(&tasks).unwrap();
-    let physical_table_id = 1024;
-    let procedure = AlterLogicalTablesProcedure::new(tasks, physical_table_id, ddl_context);
-
-    assert_event_filter_rejects(&procedure);
-    let lifecycle_state = ProcedureState::Running;
-    let event = procedure
-        .event(&EventContext {
-            procedure_id: ProcedureId::random(),
-            lifecycle_state: &lifecycle_state,
-            trigger: EventTrigger::Submitted,
-            event_type_filter: Arc::new(EventTypeFilter::All),
-        })
-        .unwrap();
-
-    assert_eq!(event.event_type(), "alter_logical_tables");
-    assert_eq!(
-        event.json_payload().unwrap(),
-        serde_json::json!({
-            "version": TABLE_DDL_PAYLOAD_VERSION,
-            "data": {
-                "kind": "alter_logical_tables",
-                "physical_table_id": physical_table_id,
-                "tasks": expected_tasks,
-            },
-        })
-    );
-    assert_eq!(
-        event
-            .extra_rows()
-            .unwrap()
-            .into_iter()
-            .map(|row| row.values)
-            .collect::<Vec<_>>(),
-        vec![
-            vec![
-                ValueData::StringValue(DEFAULT_CATALOG_NAME.to_string()).into(),
-                ValueData::StringValue(DEFAULT_SCHEMA_NAME.to_string()).into(),
-                ValueData::StringValue("table1".to_string()).into(),
-                Value::default(),
-                ValueData::U32Value(physical_table_id).into(),
-            ],
-            vec![
-                ValueData::StringValue(DEFAULT_CATALOG_NAME.to_string()).into(),
-                ValueData::StringValue(DEFAULT_SCHEMA_NAME.to_string()).into(),
-                ValueData::StringValue("table2".to_string()).into(),
-                Value::default(),
-                ValueData::U32Value(physical_table_id).into(),
-            ],
-        ]
-    );
-}
-
-#[test]
-fn test_later_events_are_lightweight_without_dynamic_success_id() {
-    let node_manager = Arc::new(MockDatanodeManager::new(()));
-    let ddl_context = new_ddl_context(node_manager);
-    let procedure = AlterLogicalTablesProcedure::new(
-        vec![make_alter_logical_table_add_column_task(
-            Some(DEFAULT_SCHEMA_NAME),
-            "table1",
-            vec!["column1".to_string()],
-        )],
-        1024,
-        ddl_context,
-    );
-    let lifecycle_state = ProcedureState::Running;
-    let child_id = ProcedureId::random();
-    let triggers = vec![
-        EventTrigger::Recovered,
-        EventTrigger::ChildSubmitted {
-            procedure_id: child_id,
-            outcome: ChildSubmissionOutcome::Accepted,
-        },
-        EventTrigger::Retrying {
-            phase: RetryPhase::Execute,
-            attempt: 1,
-        },
-        EventTrigger::RollingBack,
-        EventTrigger::Succeeded,
-        EventTrigger::Failed,
-        EventTrigger::Poisoned,
-    ];
-
-    for trigger in triggers {
-        let event = procedure
-            .event(&EventContext {
-                procedure_id: ProcedureId::random(),
-                lifecycle_state: &lifecycle_state,
-                trigger,
-                event_type_filter: Arc::new(EventTypeFilter::All),
-            })
-            .unwrap();
-        assert_eq!(event.event_type(), "alter_logical_tables");
-        assert_eq!(event.json_payload().unwrap(), serde_json::Value::Null);
-        assert_eq!(
-            event.extra_rows().unwrap()[0].values,
-            vec![Value::default(); 5]
         );
     }
 }
@@ -766,19 +641,5 @@ async fn test_on_submit_alter_region_request() {
             })),
             ..
         })
-    );
-}
-
-fn assert_event_filter_rejects(procedure: &dyn Procedure) {
-    let state = ProcedureState::Running;
-    assert!(
-        procedure
-            .event(&EventContext {
-                procedure_id: ProcedureId::random(),
-                lifecycle_state: &state,
-                trigger: EventTrigger::Submitted,
-                event_type_filter: Arc::new(EventTypeFilter::Only(HashSet::new())),
-            })
-            .is_none()
     );
 }
